Prevents multiple click to dial buttons showing up for a given number
Sugar is inconsistent with it's use of class=phone.  In Employees list view it has it but doesn't on Contacts.  When i added the new ListView.tpl fix for contacts subpanels I then found two click to dial buttons on employees list view.  This fix will avoid this issue from ever happening.