Fix for Issue #25, Calls are related to contacts created after call ends.
Prior to this fix, asteriskLogger is where all contacts were related to the call (except in the multi contact radio button cases)...
Now, callListener will set the contacts for a call whenever there is a single matching contact.  So, this allows contacts to be created after the call is over and still get related to the call.
Also, changed phone number length from 7 -> 10... going to make this a configurable parameter shortly.